requirements.txt
feedparser
kafka-python
pyspark==3.5.1

pandas
numpy==1.24.4

streamlit
plotly

transformers
torch
spacy==3.7.5
h5py>=3.10

beautifulsoup4
requests
schedule

psycopg2-binary